requirements.txt
openai==1.60.2
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6